refactor(bot): move title comparison out of utils into services
TitleComparisonService holds `private readonly cache: Map<string,
ArtistTitle>`, so it is a stateful service, not a util. Last of the
five classes #1991 enumerated.

Eleven references across three shapes, all rewritten: two from
queueDisplay, eight from trackUtils, and one jest.mock() literal in the
tests/ tree, which is the second source tree a src-only scan misses.

The eight from trackUtils collapse back to '../titleComparison'. They
were relative siblings before #1968 split the tree, became
'../../../utils/music/titleComparison' when trackUtils moved to
services in the slice before this one, and are siblings again now that
both live under services/musicManagement.

One outward dependency needed the reverse rewrite, since the moved code
now sits one level deeper relative to utils: '../../misc/stringUtils'
becomes '../../../utils/misc/stringUtils', in service.ts and in the
jest.mock() literal in service.spec.ts.

Closes #1991
